refactor: move AuthConstant to global.constant package
