+++ src/main/java/io/github/teamwaff1e/waffle/global/interceptor/LoginInterceptor.java
@@ -1,6 +1,6 @@
 package io.github.teamwaff1e.waffle.global.interceptor;
 
-import io.github.teamwaff1e.waffle.domain.auth.constant.AuthConstant;
+import io.github.teamwaff1e.waffle.global.constant.auth.AuthConstant;
 import io.github.teamwaff1e.waffle.global.exception.path.ErrorPath;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
